requirements.txt
flask
yt-dlp
openai-whisper
torch
numpy
transformers
accelerate
flask-socketio
sentencepiece
sacremoses
requests
